+++ SpehsEngine/Core/StringOperations.h
@@ -16,7 +16,7 @@ namespace se
 	std::string toString(const bool value);
 
 	// Extended std::to_string() function. Crops out extra precision given by std::to_string(). Maximum precision 6.
-	std::string toString(const float value, const int precision);
+	std::string toString(const float value, const size_t precision);
 
 	// If value is larger than zero, adds the '+' character prefix into the string
 	template<typename T>
@@ -27,7 +27,7 @@ namespace se
 		else
 			return std::to_string(value);
 	}
-	std::string toSignedString(const float value, const int precision);
+	std::string toSignedString(const float value, const size_t precision);
 
 	/*
 		Returns a string representing the integer in hex format.
